// packages/ra-core/src/types.ts
export type Identifier = string | number;
export interface Record {
    id: Identifier;
    [key: string]: any;
}

export interface RecordMap<RecordType = Record> {
    // Accept strings and numbers as identifiers
    [id: string]: RecordType;
    [id: number]: RecordType;
}

export interface Sort {
    field: string;
    order: string;
}
export interface Filter {
    [k: string]: any;
}
export interface Pagination {
    page: number;
    perPage: number;
}
export type ValidUntil = Date;
// ...
export type DataProvider = {
    getList: (
        resource: string,
        params: GetListParams
    ) => // params: object
    Promise<GetListResult>;

    getOne: (resource: string, params: GetOneParams) => Promise<GetOneResult>;

    getMany: (resource: string, params: GetManyParams) => Promise<GetManyResult>;

    getManyReference: (resource: string, params: GetManyReferenceParams) => Promise<GetManyReferenceResult>;

    update: (resource: string, params: UpdateParams) => Promise<UpdateResult>;

    updateMany: (resource: string, params: UpdateManyParams) => Promise<UpdateManyResult>;

    create: (resource: string, params: CreateParams) => Promise<CreateResult>;

    delete: (resource: string, params: DeleteParams) => Promise<DeleteResult>;

    deleteMany: (resource: string, params: DeleteManyParams) => Promise<DeleteManyResult>;
    //TODO: hyojin
    //=============================custom==========================
    createForUploadFormData: (resource: string, params: CreateForUploadFormDataParams) => Promise<CreateResult>;
    deleteWithBody: (resource: string, body: any) => Promise<DeleteResult>;
    updateSetMethod: (resource: string, params: UpdateSetMethodParams) => Promise<UpdateResult>;
    //=============================custom==========================

    [key: string]: any;
};

export interface GetListParams {
    pagination: Pagination;
    sort: Sort;
    filter: any;
}
export interface GetListResult {
    data: Record[];
    total: number;
    validUntil?: ValidUntil;
}

export interface GetOneParams {
    id: Identifier;
}
export interface GetOneResult {
    data: Record;
    validUntil?: ValidUntil;
}

export interface GetManyParams {
    ids: Identifier[];
}
export interface GetManyResult {
    data: Record[];
    validUntil?: ValidUntil;
}

export interface GetManyReferenceParams {
    target: string;
    id: Identifier;
    pagination: Pagination;
    sort: Sort;
    filter: any;
}
export interface GetManyReferenceResult {
    data: Record[];
    total: number;
    validUntil?: ValidUntil;
}

export interface UpdateParams {
    id: Identifier;
    data: any;
    previousData: Record;
}

export interface UpdateResult {
    data: Record;
    validUntil?: ValidUntil;
}

export interface UpdateManyParams {
    ids: Identifier[];
    data: any;
}
export interface UpdateManyResult {
    data?: Identifier[];
    validUntil?: ValidUntil;
}

export interface CreateParams {
    data: any;
}
export interface CreateResult {
    data: Record;
    validUntil?: ValidUntil;
}

export interface DeleteParams {
    id: Identifier;
}
export interface DeleteResult {
    data?: Record;
}

export interface DeleteManyParams {
    ids: Identifier[];
}
export interface DeleteManyResult {
    data?: Identifier[];
}
//TODO: hyojin
//=============================custom==========================
export interface UpdateSetMethodParams {
    id: Identifier;
    data: any;
    method: string;
    previousData: Record;
}
export interface CreateForUploadFormDataParams {
    formData: FormData;
}
// ...
export type LegacyDataProvider = (type: string, resource: string, params: any) => Promise<any>;